+++ src/Sulu/Bundle/SearchBundle/Tests/Unit/Controller/WebsiteSearchControllerTest.php
@@ -19,6 +19,7 @@ use Sulu\Component\Localization\Localization;
 use Sulu\Component\Webspace\Analyzer\RequestAnalyzerInterface;
 use Sulu\Component\Webspace\Webspace;
 use Symfony\Bundle\FrameworkBundle\Templating\EngineInterface;
+use Symfony\Component\DependencyInjection\ContainerInterface;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 
@@ -43,6 +44,10 @@ class WebsiteSearchControllerTest extends \PHPUnit_Framework_TestCase
      * @var EngineInterface
      */
     private $engine;
+    /**
+     * @var ContainerInterface
+     */
+    private $container;
 
     /**
      * @var WebsiteSearchController
@@ -55,6 +60,7 @@ class WebsiteSearchControllerTest extends \PHPUnit_Framework_TestCase
         $this->requestAnalyzer = $this->prophesize(RequestAnalyzerInterface::class);
         $this->parameterResolver = $this->prophesize(ParameterResolverInterface::class);
         $this->engine = $this->prophesize(EngineInterface::class);
+        $this->container = $this->prophesize(ContainerInterface::class);
 
         $this->websiteSearchController = new WebsiteSearchController(
             $this->searchManager->reveal(),
@@ -62,6 +68,7 @@ class WebsiteSearchControllerTest extends \PHPUnit_Framework_TestCase
             $this->parameterResolver->reveal(),
             $this->engine->reveal()
         );
+        $this->websiteSearchController->setContainer($this->container->reveal());
     }
 
     public function testQueryAction()
@@ -78,10 +85,19 @@ class WebsiteSearchControllerTest extends \PHPUnit_Framework_TestCase
         $this->requestAnalyzer->getCurrentLocalization()->willReturn($localization);
         $this->requestAnalyzer->getWebspace()->willReturn($webspace);
 
+        $this->container->getParameter('sulu_search.website_indexes')->willReturn(['page_{ webspace_key }_published']);
+
+        /** @var \Twig_Environment $twigEnvironment */
+        $twigEnvironment = $this->prophesize(\Twig_Environment::class);
+        $this->container->get('twig')->willReturn($twigEnvironment->reveal());
+        $twigTemplate = $this->prophesize(\Twig_Template::class);
+        $twigEnvironment->createTemplate('page_{ webspace_key }_published')->willReturn($twigTemplate);
+        $twigTemplate->render(['webspace' => $webspace])->willReturn('page_sulu_published');
+
         $searchQueryBuilder = $this->prophesize(SearchQueryBuilder::class);
         $this->searchManager->createSearch('+("Test" OR Test* OR Test~) ')->willReturn($searchQueryBuilder->reveal());
         $searchQueryBuilder->locale('en')->willReturn($searchQueryBuilder->reveal());
-        $searchQueryBuilder->index('page_sulu_published')->willReturn($searchQueryBuilder->reveal());
+        $searchQueryBuilder->indexes(['page_sulu_published'])->willReturn($searchQueryBuilder->reveal());
         $searchQueryBuilder->execute()->willReturn([]);
 
         $this->parameterResolver->resolve(